Update fakku.py
added a command line parameter to allow the user to set a different value for the amount of retry attempts. 
added 2 variables for the Default amount of attempts and the Max amount of attempts 
Save now returns a Boolean if non 0 (or True) download success false otherwise. ( in case the download fail more than the amount expected of times. )